setup/requirements.txt
yfinance>=0.2.36
pandas>=2.1.0
PyYAML>=6.0.1